+++ common/src/main/java/dev/tocraft/remorphed/command/RemorphedCommand.java
@@ -362,6 +362,7 @@ public class RemorphedCommand implements CommandEvents.CommandRegistration {
             rootNode.addChild(hasSkin);
         }
 
+
         dispatcher.getRoot().addChild(rootNode);
 
     }