+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+remorphed 7.1.1
+================
+
+- optimize menu rendering by FugLong
+
 remorphed 7.1.0
 ================
 
@@ -5,7 +10,6 @@ remorphed 7.1.0
 - fix broken counter in menu
 - add delete functionality to Menu - press "X" or mouse-scroll-wheel to delete
 - fix duplicate entities in menu
-- optimize menu rendering by FugLong
 
 remorphed 7.0
 ================